add method

+++ hutool-crypto/src/main/java/cn/hutool/crypto/digest/BCrypt.java
@@ -179,7 +179,7 @@ public class BCrypt {
 	private static byte char64(char x) {
 		if ((int) x > index_64.length)
 			return -1;
-		return index_64[(int) x];
+		return index_64[x];
 	}
 
 	/**
